requirements.txt
aniso8601==3.0.0
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.2
django-allauth==0.36.0
djangorestframework==3.8.2
graphene==2.1.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
gunicorn==19.8.1
idna==2.6
iso8601==0.1.12
oauthlib==2.1.0
promise==2.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.4
urllib3==1.22
virtualenv==16.0.0
whitenoise==3.3.1
